# backend/docs/swagger.yaml
basePath: /
definitions:
  models.Button:
    properties:
      action:
        example: create
        type: string
      code:
        example: create
        type: string
      created_at:
        type: string
      id:
        example: 1
        type: integer
      menu_id:
        example: 1
        type: integer
      name:
        example: 创建用户
        type: string
      permission_code:
        example: user:create
        type: string
      updated_at:
        type: string
    type: object
...
  models.Permission:
    properties:
      button_id:
        example: 1
        type: integer
      created_at:
        type: string
      id:
        example: 1
        type: integer
      menu_id:
        example: 1
        type: integer
      name:
        example: 查看仪表盘
        type: string
      permission_code:
        example: view_dashboard
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      type:
        example: menu
        type: string
      updated_at:
        type: string
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      description:
        example: 系统管理员角色
        type: string
      id:
        example: 1
        type: integer
      name:
        example: 管理员
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.User:
    properties:
      created_at:
        type: string
      id:
        example: 1
        type: integer
      password:
        example: password123
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
      username:
        example: admin
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 租户中心系统的API文档
  title: 租户中心API
  version: "1.0"
paths:
  /api/buttons:
    post:
      consumes:
      - application/json
      description: 创建新按钮，需要管理员权限
      parameters:
      - description: 按钮信息
        in: body
        name: button
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: 按钮创建成功
          schema:
            type: object
        "400":
          description: 无效的请求参数
          schema:
            type: object
        "500":
          description: 创建按钮失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 按钮管理
  /api/buttons/{id}:
    put:
      consumes:
      - application/json
      description: 更新指定按钮的信息，需要管理员权限
      parameters:
      - description: 按钮ID
        in: path
        name: id
        required: true
        type: integer
      - description: 按钮信息
        in: body
        name: button
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 按钮信息更新成功
          schema:
            type: object
        "400":
          description: 无效的请求参数
          schema:
            type: object
        "500":
          description: 更新按钮信息失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 按钮管理
  /api/buttons/{id}/permission:
    post:
      consumes:
      - application/json
      description: 为指定按钮绑定权限，需要管理员权限
      parameters:
      - description: 按钮ID
        in: path
        name: id
        required: true
        type: integer
      - description: 权限信息
        in: body
        name: permission
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 权限绑定成功
          schema:
            type: object
        "400":
          description: 无效的请求参数
          schema:
            type: object
        "500":
          description: 绑定权限失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 按钮管理
  /api/buttons/{id}/permissions:
    get:
      consumes:
      - application/json
      description: 获取指定按钮的权限列表，需要管理员权限
      parameters:
      - description: 按钮ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 权限列表
          schema:
            items:
              $ref: '#/definitions/models.Permission'
            type: array
        "400":
          description: 无效的按钮ID
          schema:
            type: object
        "500":
          description: 获取按钮权限列表失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 按钮管理
  /api/buttons/menu/{menuId}:
    get:
      consumes:
      - application/json
      description: 获取指定菜单的所有按钮列表，需要管理员权限
      parameters:
      - description: 菜单ID
        in: path
        name: menuId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 按钮列表
          schema:
            items:
              $ref: '#/definitions/models.Button'
            type: array
        "400":
          description: 无效的菜单ID
          schema:
            type: object
        "500":
          description: 获取按钮列表失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 按钮管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"